kernel-c18n: compile an IR file of the kernel
The resulting IR file can be used to analyse a control flow graph for
compartmentalization.

The build system is modified to specify stage-specific flags in a
separate CFLAGS_STAGE make variable that is usually set to '-c -o
${.TARGET}' but can be overwritten with '-S -emit-llvm -o ${.TARGET}'
at the make target level. The IR file is built when the COMPILE_IR make
option is defined in a kernel configuration.

The kernel in the IR output is linked using *.llo files corresponding to
*.o files. At the moment, only C files are compiled into IR and assembly
(.S) files are compiled into empty files. In the future, we can try to
translate assembly files into IR files that include information relevent
to the analysis. For example, assembly macros for branching could be
translated into function calls in IR.

The changes in config(8) are minimal but should be refactored in case
we'd like to upstream them. They are structured to be minimal not to
create conflicts with upstream and hence they are not easy to follow.